requirement.txt
requests>=2.28
PyYAML>=6.0
pandas>=2.0
pydantic>=1.10
tqdm>=4.0
jsonlines>=2.0
matplotlib>=3.0
openai>=1.0 
python-dotenv>=1.0